+++ solvers/src/muoblpsolvers/mes/mes_add1.py
@@ -5,10 +5,9 @@ from muoblp.model.multi_objective_lp import MultiObjectiveLpProblem
 from muoblpbindings import equal_shares_add1
 from pulp import LpSolver
 
-from muoblpsolvers.common import (
-    prepare_mes_parameters,
-    set_selected_candidates,
-)
+from muoblpsolvers.utils import set_solved
+
+from .common import prepare_mes_parameters
 
 logger = logging.getLogger(__name__)
 
@@ -39,6 +38,8 @@ class MethodOfEqualSharesAdd1Solver(LpSolver):
             total_utilities,
             total_budget,
         )
-        set_selected_candidates(lp, selected)
 
         logger.info("SOLVER END", extra={"time": time.time() - start_time})
+
+        set_solved(lp, selected)
+        return lp.status
