+++ solvers/src/muoblpsolvers/mes/mes_constrains.py
@@ -6,10 +6,7 @@ from muoblp.model.multi_objective_lp import MultiObjectiveLpProblem
 from muoblpbindings import equal_shares_utils
 from pulp import LpConstraint, LpConstraintGE, LpConstraintLE, LpSolver
 
-from muoblpsolvers.common import (
-    prepare_mes_parameters,
-    set_selected_candidates,
-)
+from .common import prepare_mes_parameters
 
 logger = logging.getLogger(__name__)
 
@@ -84,7 +81,8 @@ class MethodOfEqualSharesConstrainsSolver(LpSolver):
                 total_budget,
             )
             logger.debug(f"FINISHED MES {time.time() - start_time:.2f} s\n")
-            set_selected_candidates(lp, selected)
+            for variable in lp.variables():
+                variable.setInitialValue(1 if variable.name in selected else 0)
 
             # Check constraints
             infeasible = get_infeasible_constraints(lp)
